Moved elements processors into their own directory.
