+++ lib/Byng/Pimcore/ElasticSearch/Processor/Page/PageProcessorFactory.php
@@ -15,6 +15,9 @@ namespace Byng\Pimcore\Elasticsearch\Processor\Page;
 
 use Byng\Pimcore\Elasticsearch\Filter\FilterInterface;
 use Byng\Pimcore\Elasticsearch\Filter\TagKeyFilter;
+use Byng\Pimcore\Elasticsearch\Processor\Element\DateElementProcessor;
+use Byng\Pimcore\Elasticsearch\Processor\Element\ElementProcessor;
+use Byng\Pimcore\Elasticsearch\Processor\Element\SelectElementProcessor;
 use NF\HtmlToText;
 
 /**
